Season policy

+++ modules/Abd/Course/Http/Controllers/SeasonController.php
@@ -5,6 +5,7 @@ namespace Abd\Course\Http\Controllers;
 use Abd\Common\Responses\AjaxResponses;
 use Abd\Course\Http\Requests\SeasonRequest;
 use Abd\Course\Models\Season;
+use Abd\Course\Repositories\CourseRepo;
 use Abd\Course\Repositories\SeasonRepo;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
@@ -17,8 +18,9 @@ class SeasonController extends Controller
         $this->seasonRepo = $seasonRepo;
     }
 
-    public function store($course, SeasonRequest $request)
+    public function store($course, SeasonRequest $request, CourseRepo $courseRepo)
     {
+        $this->authorize('createSeason', $courseRepo->findById($course));
         $this->seasonRepo->store($course, $request);
         newFeedback();
         return back();
@@ -27,11 +29,13 @@ class SeasonController extends Controller
     public function edit($id)
     {
         $season = $this->seasonRepo->findById($id);
+        $this->authorize('edit', $season);
         return view('Courses::seasons.edit', compact('season'));
     }
 
     public function update($id, SeasonRequest $request)
     {
+        $this->authorize('edit', $this->seasonRepo->findById($id));
         $this->seasonRepo->update($id, $request);
         newFeedback();
         return back();
@@ -40,12 +44,14 @@ class SeasonController extends Controller
     public function destroy($id)
     {
         $season = $this->seasonRepo->findById($id);
+        $this->authorize('delete', $season);
         $season->delete();
         return AjaxResponses::SuccessResponse();
     }
 
     public function accept($id)
     {
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateConfirmationStatus($id, Season::CONFIRMATION_STATUS_ACCEPTED)) {
             return AjaxResponses::SuccessResponse();
         } else {
@@ -55,6 +61,7 @@ class SeasonController extends Controller
 
     public function reject($id)
     {
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateConfirmationStatus($id, Season::CONFIRMATION_STATUS_REJECTED)) {
             return AjaxResponses::SuccessResponse();
         } else {
@@ -64,6 +71,7 @@ class SeasonController extends Controller
 
     public function lock($id)
     {
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateStatus($id, Season::STATUS_LOCKED)) {
             return AjaxResponses::SuccessResponse();
         } else {
@@ -73,6 +81,7 @@ class SeasonController extends Controller
 
     public function unlock($id)
     {
+        $this->authorize('change_confirmation_status', Season::class);
         if ($this->seasonRepo->updateStatus($id, Season::STATUS_OPENED)) {
             return AjaxResponses::SuccessResponse();
         } else {
